Move chartData calculation to outer layer of code
Only the entrypoints to inner modules now calculate the chart data
