+++ src/viewer.js
@@ -3,7 +3,6 @@ const fs = require('fs');
 const http = require('http');
 
 const WebSocket = require('ws');
-const _ = require('lodash');
 const express = require('express');
 const ejs = require('ejs');
 const opener = require('opener');
@@ -11,7 +10,6 @@ const mkdir = require('mkdirp');
 const { bold } = require('chalk');
 
 const Logger = require('./Logger');
-const analyzer = require('./analyzer');
 
 const projectRoot = path.resolve(__dirname, '..');
 
@@ -22,17 +20,15 @@ module.exports = {
   start: startServer
 };
 
-async function startServer(bundleStats, opts) {
+async function startServer(chartData, opts) {
   const {
     port = 8888,
     host = '127.0.0.1',
     openBrowser = true,
-    bundleDir = null,
+    // bundleDir = null,
     logger = new Logger(),
     defaultSizes = 'parsed'
   } = opts || {};
-
-  let chartData = getChartData(logger, bundleStats, bundleDir);
 
   if (!chartData) return;
 
@@ -80,9 +76,7 @@ async function startServer(bundleStats, opts) {
     updateChartData
   };
 
-  function updateChartData(bundleStats) {
-    const newChartData = getChartData(logger, bundleStats, bundleDir);
-
+  function updateChartData(newChartData) {
     if (!newChartData) return;
 
     chartData = newChartData;
@@ -98,7 +92,7 @@ async function startServer(bundleStats, opts) {
   }
 }
 
-function generateReport(bundleStats, opts) {
+function generateReport(chartData, opts) {
   const {
     openBrowser = true,
     reportFilename = 'report.html',
@@ -106,8 +100,6 @@ function generateReport(bundleStats, opts) {
     logger = new Logger(),
     defaultSizes = 'parsed'
   } = opts || {};
-
-  const chartData = getChartData(logger, bundleStats, bundleDir);
 
   if (!chartData) return;
 
@@ -140,22 +132,4 @@ function generateReport(bundleStats, opts) {
 
 function getAssetContent(filename) {
   return fs.readFileSync(`${projectRoot}/public/${filename}`, 'utf8');
-}
-
-function getChartData(logger, ...args) {
-  let chartData;
-
-  try {
-    chartData = analyzer.getViewerData(...args, { logger });
-  } catch (err) {
-    logger.error(`Could't analyze webpack bundle:\n${err}`);
-    chartData = null;
-  }
-
-  if (_.isEmpty(chartData)) {
-    logger.error("Could't find any javascript bundles in provided stats file");
-    chartData = null;
-  }
-
-  return chartData;
 }
